+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Extract geojson feature parsing into [Feature Parser](https://github.com/dmfenton/feature-parser)
+* Use Highland 3.0 branch for more performant pipelines
+
 ## [1.0.14] - 2016-01-28
 ### Fixed
 * Rebuild